Move WebRTC configuration to signaling server
