+++ pkg/app/config/config.go
@@ -22,13 +22,10 @@ import (
 	"github.com/google/cloud-android-orchestration/pkg/app/encryption"
 	"github.com/google/cloud-android-orchestration/pkg/app/instances"
 	"github.com/google/cloud-android-orchestration/pkg/app/secrets"
+	"github.com/google/cloud-android-orchestration/pkg/app/signaling"
 
 	toml "github.com/pelletier/go-toml"
 )
-
-type WebRTCConfig struct {
-	STUNServers []string
-}
 
 type Config struct {
 	WebStaticFilesPath string
@@ -37,7 +34,7 @@ type Config struct {
 	InstanceManager    instances.Config
 	EncryptionService  encryption.Config
 	DatabaseService    database.Config
-	WebRTC             WebRTCConfig
+	WebRTC             signaling.WebRTCConfig
 }
 
 const DefaultConfFile = "conf.toml"
